backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.9.5
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
dateparser==1.2.2
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.7
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.6.2
numpy==1.26.4
orjson==3.10.18
pandas==2.2.2
pandas_ta==0.3.14b0
propcache==0.3.2
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
python-binance==1.0.19
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rich-toolkit==0.14.7
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.37.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
ujson==5.10.0
urllib3==2.5.0
uvicorn==0.29.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==12.0
yarl==1.20.1
